+++ discO/plugin/galpy/wrapper.py
@@ -269,6 +269,20 @@ class GalpyPotentialWrapper(
 
     """
 
+    def __init__(
+        self,
+        potential: T.Any,
+        *,
+        frame: TH.OptFrameLikeType = None,
+        representation_type: TH.OptRepresentationLikeType = None,
+    ):
+        super().__init__(
+            potential,
+            frame=frame,
+            representation_type=representation_type,
+        )
+        getattr(self.wrapped, "turn_physical_on", object)()
+
 
 # /class